diar/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pyannote.audio==3.1.1
soundfile==0.12.1
numpy==1.26.4
httpx==0.27.0
av==12.1.0
python-multipart==0.0.9
